Use guint8 to store color components rather than gdouble.
Color components in TTML are 8 bits each.

+++ libs/subtitle/gstsubtitle.h
@@ -72,10 +72,10 @@ typedef enum {
 } GstSubtitleOverflowMode;
 
 struct _GstSubtitleColor {
-  gdouble r;
-  gdouble g;
-  gdouble b;
-  gdouble a;
+  guint8 r;
+  guint8 g;
+  guint8 b;
+  guint8 a;
 };
 
 typedef enum {
